TIP-1160: Split into 2 scripts
