+++ upgrades/schema/Version_4_0_20190916122239_remove_product_empty_raw_values.php
@@ -3,7 +3,6 @@
 namespace Pim\Upgrade\Schema;
 
 use Akeneo\Pim\Enrichment\Bundle\Elasticsearch\Indexer\ProductIndexer;
-use Akeneo\Pim\Enrichment\Bundle\Elasticsearch\Indexer\ProductModelIndexer;
 use Akeneo\Pim\Enrichment\Component\Product\Factory\EmptyValuesCleaner;
 use Doctrine\DBAL\Schema\Schema;
 use Doctrine\DBAL\Types\Type;
@@ -12,15 +11,15 @@ use Symfony\Component\DependencyInjection\ContainerAwareInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
- * This migration will delete the empty values from raw values of product and product models.
+ * This migration will delete the empty values from raw values of products.
  * For example, the value {attr: {<all_channels>: {<all_locales: []}}} will be removed from the raw_values field.
  */
-final class Version_4_0_20190916122239_remove_empty_raw_values extends AbstractMigration implements ContainerAwareInterface
+final class Version_4_0_20190916122239_remove_product_empty_raw_values extends AbstractMigration implements ContainerAwareInterface
 {
     /** @var ContainerInterface */
     private $container;
 
-    const BATCH_SIZE = 100;
+    private const BATCH_SIZE = 1000;
 
     /**
      * {@inheritdoc}
@@ -33,7 +32,6 @@ final class Version_4_0_20190916122239_remove_empty_raw_values extends AbstractM
     public function up(Schema $schema) : void
     {
         $this->cleanProducts();
-        $this->cleanProductModels();
     }
 
     private function cleanProducts()
@@ -79,49 +77,6 @@ final class Version_4_0_20190916122239_remove_empty_raw_values extends AbstractM
         $this->getProductIndexer()->indexFromProductIdentifiers($productIdentifiersToIndex);
     }
 
-    private function cleanProductModels()
-    {
-        $productModelsToProcess = true;
-        $productModelCodesToIndex = [];
-        $page = 0;
-        while ($productModelsToProcess) {
-            $productModelsToProcess = false;
-            $sql = sprintf(
-                "SELECT code, raw_values FROM pim_catalog_product_model LIMIT %d, %s",
-                $page * self::BATCH_SIZE,
-                self::BATCH_SIZE
-            );
-            $rows = $this->connection->executeQuery($sql)->fetchAll();
-
-            foreach ($rows as $row) {
-                $productModelsToProcess = true;
-                $rawValues = json_decode($row['raw_values'], true);
-                $cleanRawValues = $this->getValueCleaner()->cleanAllValues(['ID' => $rawValues])['ID'];
-                if ($rawValues !== $cleanRawValues) {
-                    $this->connection->executeQuery(
-                        'UPDATE pim_catalog_product_model SET raw_values = :rawValues WHERE code = :code',
-                        [
-                            'rawValues' => json_encode($cleanRawValues),
-                            'code' => $row['code']
-                        ], [
-                            'rawValues' => Type::STRING,
-                            'code' => Type::STRING
-                        ]
-                    );
-                    $productModelCodesToIndex[] = $row['code'];
-                    if (count($productModelCodesToIndex) % self::BATCH_SIZE === 0) {
-                        $this->getProductModelIndexer()->indexFromProductModelCodes($productModelCodesToIndex);
-                        $productModelCodesToIndex = [];
-                    }
-                }
-            }
-
-            $page++;
-        }
-
-        $this->getProductModelIndexer()->indexFromProductModelCodes($productModelCodesToIndex);
-    }
-
     public function down(Schema $schema) : void
     {
     }
@@ -134,10 +89,5 @@ final class Version_4_0_20190916122239_remove_empty_raw_values extends AbstractM
     private function getProductIndexer(): ProductIndexer
     {
         return $this->container->get('pim_catalog.elasticsearch.indexer.product');
-    }
-
-    private function getProductModelIndexer(): ProductModelIndexer
-    {
-        return $this->container->get('pim_catalog.elasticsearch.indexer.product_model');
     }
 }
